经验代码

+++ _posts/DDKJ/11、经验代码/2、springBoot模板项目/1_Mapper接口/hlj-demo/hlj-data/src/main/java/com/hlj/proj/data/common/query/PagingQuery.java
@@ -21,7 +21,7 @@ public abstract class PagingQuery implements Serializable{
 	/**
 	 * 数据记录总数
 	 */
-	private Integer itemCount;
+	private Long itemCount;
 	private Integer startRow;
 	private Integer endRow;
 	
@@ -31,7 +31,7 @@ public abstract class PagingQuery implements Serializable{
 	}
 
 
-    public void setItemCount(Integer itemCount){
+    public void setItemCount(Long itemCount){
     	this.itemCount = itemCount;
     	startRow = (pageNo - 1) * pageSize;
     	endRow = pageSize;
